# Conflicts:
#	backend/base/models.py
#	backend/building_comment/views.py
#	backend/email_template/tests.py

+++ backend/building/views.py
@@ -1,21 +1,22 @@
+from drf_spectacular.utils import extend_schema
 from rest_framework import permissions
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
 
+from base.permissions import ReadOnlyOwnerOfBuilding, IsAdmin, IsSuperStudent, ReadOnlyStudent
 from base.models import Building
 from base.serializers import BuildingSerializer
 from util.request_response_util import *
-from drf_spectacular.utils import extend_schema
 
-# TODO:  we don't actually have to work with 'syndic' key, we can also require 'syndic_id' as parameter in body
-#  however, de automatic documentation might be a bit harder?
+
 TRANSLATE = {"syndic": "syndic_id"}
 
 
 class DefaultBuilding(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsAdmin | IsSuperStudent]
     serializer_class = BuildingSerializer
 
-    @extend_schema(responses={201: BuildingSerializer, 400: None})
+    @extend_schema(responses=post_docs(BuildingSerializer))
     def post(self, request):
         """
         Create a new building
@@ -34,10 +35,10 @@ class DefaultBuilding(APIView):
 
 
 class BuildingIndividualView(APIView):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsAdmin | IsSuperStudent | ReadOnlyStudent | ReadOnlyOwnerOfBuilding]
     serializer_class = BuildingSerializer
 
-    @extend_schema(responses={200: BuildingSerializer, 400: None})
+    @extend_schema(responses=get_docs(BuildingSerializer))
     def get(self, request, building_id):
         """
         Get info about building with given id
@@ -48,10 +49,11 @@ class BuildingIndividualView(APIView):
             return bad_request(object_name="Building")
 
         building_instance = building_instance[0]
+        self.check_object_permissions(request, building_instance)
         serializer = BuildingSerializer(building_instance)
         return get_success(serializer)
 
-    @extend_schema(responses={204: None, 400: None})
+    @extend_schema(responses=delete_docs())
     def delete(self, request, building_id):
         """
         Delete building with given id
@@ -61,10 +63,12 @@ class BuildingIndividualView(APIView):
             return bad_request(object_name="Building")
         building_instance = building_instance[0]
 
+        self.check_object_permissions(request, building_instance)
+
         building_instance.delete()
         return delete_success()
 
-    @extend_schema(responses={200: BuildingSerializer, 400: None})
+    @extend_schema(responses=patch_docs(BuildingSerializer))
     def patch(self, request, building_id):
         """
         Edit building with given ID
@@ -74,6 +78,7 @@ class BuildingIndividualView(APIView):
             return bad_request(object_name="Building")
 
         building_instance = building_instance[0]
+        self.check_object_permissions(request, building_instance)
         data = request_to_dict(request.data)
 
         set_keys_of_instance(building_instance, data, TRANSLATE)
@@ -84,7 +89,55 @@ class BuildingIndividualView(APIView):
         return patch_success(BuildingSerializer(building_instance))
 
 
+# The building url you can access without account
+class BuildingPublicView(APIView):
+    serializer_class = BuildingSerializer
+
+    @extend_schema(responses=get_docs(BuildingSerializer))
+    def get(self, request, building_public_id):
+        """
+        Get building with the public id
+        """
+        building_instance = Building.objects.filter(public_id=building_public_id)
+
+        if not building_instance:
+            return not_found("Building")
+
+        building_instance = building_instance[0]
+
+        # TODO: should the general public see all data about a building?
+        #  Discuss this when writing tests for building
+        return get_success(BuildingSerializer(building_instance))
+
+
+class BuildingNewPublicId(APIView):
+    serializer_class = BuildingSerializer
+
+    @extend_schema(
+        description="Generate a new unique uuid as public id for the building.",
+        responses=post_docs(BuildingSerializer),
+    )
+    def post(self, request, building_id):
+        """
+        Generate a new public_id for the building with given id
+        """
+        building_instance = Building.objects.filter(id=building_id)
+
+        if not building_instance:
+            return bad_request("Building")
+
+        building_instance = building_instance[0]
+
+        building_instance.public_id = get_unique_uuid()
+
+        if r := try_full_clean_and_save(building_instance):
+            return r
+
+        return post_success(BuildingSerializer(building_instance))
+
+
 class AllBuildingsView(APIView):
+    permission_classes = [IsAuthenticated, IsAdmin | IsSuperStudent]
     serializer_class = BuildingSerializer
 
     def get(self, request):
@@ -98,14 +151,18 @@ class AllBuildingsView(APIView):
 
 
 class BuildingOwnerView(APIView):
+    permission_classes = [IsAuthenticated, IsAdmin | IsSuperStudent | ReadOnlyOwnerOfBuilding]
     serializer_class = BuildingSerializer
 
-    @extend_schema(responses={200: BuildingSerializer, 400: None})
+    @extend_schema(responses=get_docs(BuildingSerializer))
     def get(self, request, owner_id):
         """
         Get all buildings owned by syndic with given id
         """
         building_instance = Building.objects.filter(syndic=owner_id)
+
+        for b in building_instance:
+            self.check_object_permissions(request, b)
 
         if not building_instance:
             return bad_request_relation("building", "syndic")
