+++ backend/users/views.py
@@ -1,13 +1,21 @@
 import json
 
-from rest_framework import permissions
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
 
+from base.permissions import (
+    IsAdmin,
+    IsSuperStudent,
+    OwnerAccount,
+    CanEditUser,
+    CanEditRole,
+    CanDeleteUser,
+    CanCreateUser,
+)
 from base.models import User
 from base.serializers import UserSerializer
 from util.request_response_util import *
 from drf_spectacular.utils import extend_schema
-
 
 TRANSLATE = {"role": "role_id"}
 
@@ -31,13 +39,14 @@ def _try_adding_region_to_user_instance(user_instance, region_value):
 
 
 class DefaultUser(APIView):
+    permission_classes = [IsAuthenticated, IsAdmin | IsSuperStudent, CanCreateUser]
     serializer_class = UserSerializer
-    # TODO: authorization
-    # permission_classes = [permissions.IsAuthenticated]
 
     # TODO: in order for this to work, you have to pass a password
     #  In the future, we probably won't use POST this way anymore (if we work with the whitelist method)
-    @extend_schema(responses={201: UserSerializer, 400: None})
+    #  However, an easy workaround would be to add a default value to password (in e.g. `clean`)
+    #     -> probably the easiest way
+    @extend_schema(responses=post_docs(UserSerializer))
     def post(self, request):
         """
         Create a new user
@@ -47,6 +56,8 @@ class DefaultUser(APIView):
         user_instance = User()
 
         set_keys_of_instance(user_instance, data, TRANSLATE)
+
+        self.check_object_permissions(request, user_instance)
 
         if r := try_full_clean_and_save(user_instance):
             return r
@@ -64,10 +75,16 @@ class DefaultUser(APIView):
 
 
 class UserIndividualView(APIView):
+    permission_classes = [
+        IsAuthenticated,
+        IsAdmin | IsSuperStudent | OwnerAccount,
+        CanEditUser,
+        CanEditRole,
+        CanDeleteUser,
+    ]
     serializer_class = UserSerializer
-    permission_classes = [permissions.IsAuthenticated]
 
-    @extend_schema(responses={200: UserSerializer, 400: None})
+    @extend_schema(responses=get_docs(UserSerializer))
     def get(self, request, user_id):
         """
         Get info about user with given id
@@ -76,28 +93,32 @@ class UserIndividualView(APIView):
 
         if not user_instance:
             return bad_request(object_name="User")
+        user_instance = user_instance[0]
 
-        serializer = UserSerializer(user_instance[0])
+        self.check_object_permissions(request, user_instance)
+
+        serializer = UserSerializer(user_instance)
         return get_success(serializer)
 
-    @extend_schema(responses={204: None, 400: None})
+    @extend_schema(responses=delete_docs())
     def delete(self, request, user_id):
         """
         Delete user with given id
-        We don't acutally delete a user, we put the user on inactive mode
+        We don't actually delete a user, we put the user on inactive mode
         """
         user_instance = User.objects.filter(id=user_id)
         if not user_instance:
             return bad_request(object_name="User")
-
         user_instance = user_instance[0]
+
+        self.check_object_permissions(request, user_instance)
 
         user_instance.is_active = False
         user_instance.save()
 
         return delete_success()
 
-    @extend_schema(responses={200: UserSerializer, 400: None})
+    @extend_schema(responses=patch_docs(UserSerializer))
     def patch(self, request, user_id):
         """
         Edit user with given id
@@ -108,6 +129,8 @@ class UserIndividualView(APIView):
             return bad_request(object_name="User")
 
         user_instance = user_instance[0]
+
+        self.check_object_permissions(request, user_instance)
 
         data = request_to_dict(request.data)
 
@@ -129,6 +152,7 @@ class UserIndividualView(APIView):
 
 
 class AllUsersView(APIView):
+    permission_classes = [IsAuthenticated, IsAdmin | IsSuperStudent]
     serializer_class = UserSerializer
 
     def get(self, request):
